revisar

+++ src/app/main/features/content/mantenimiento/proveedor-agencia/proveedor-agencia.component.ts
@@ -3,6 +3,9 @@ import {FormControl, FormGroup} from "@angular/forms";
 import {ProveedorAgenciaService} from "../../../../../shared/services/api/maintenence/proveedor-agencia-service";
 import {translateType} from "../../../../../shared/models/documentation/type.model";
 import {NgbModal} from "@ng-bootstrap/ng-bootstrap";
+import {CsvService} from "../../../../../shared/services/csv/csv.Service";
+import {SwalService} from "../../../../../shared/services/swal/swal.service";
+import {TranslateService} from "@ngx-translate/core";
 
 
 @Component({
@@ -19,23 +22,29 @@ export class ProveedorAgenciaComponent {
   private historial:string = "";
   protected counter:number = 3;
   protected providerCreator:boolean=false;
+  protected providerFilter:boolean=true;
   protected itemList:any[]=[];
   protected asc:boolean = true;
   private exportFiltro = {nombre: ""};
   private modalRef :any;
   protected editName:any[] = [];
 
+  protected pageSize:number=5;
+  protected currentPage:number=1;
 
   constructor(private _fetch: ProveedorAgenciaService,
-              private _modal: NgbModal) {
+              private _modal: NgbModal,
+              private  _csvService: CsvService,
+              private _swal: SwalService,
+              private _translate:TranslateService) {
     this.loadForms();
-    this.getProveedores('1');
+    this.getProveedores(1);
   }
   private loadForms(){
     this.fecthForm = new FormGroup({
       nombre: new FormControl<string>(""),
       datos_paginacion: new FormGroup({
-        pagina: new FormControl<string>("1"),
+        pagina: new FormControl<number>(1),
         num_resultados: new FormControl<string>("5"),
         orden: new FormControl<string>("nombre"),
         tipo_orden: new FormControl<string>("ASC"),
@@ -55,8 +64,9 @@ export class ProveedorAgenciaComponent {
 
 
   protected submitSearch(form:any) {
+    this.pageSize = this.fecthForm.get("datos_paginacion")?.get("num_resultados")?.value;
     if (form.status === 'VALID') {
-      this.getProveedores('1');
+      this.getProveedores(1);
     } else {
 
       //TODO:notification
@@ -69,11 +79,11 @@ export class ProveedorAgenciaComponent {
     }
   };
 
-  protected getProveedores(pagina:string | undefined){
+  protected getProveedores(pagina:number | undefined){
     this.providerCreator = false;
     // crmLoadingPage(true);
     if (pagina !== undefined) {
-      //igualar paginacion a la pagina
+      this.fecthForm.patchValue({ datos_paginacion: {pagina:this.currentPage}});
     }
 
     this._fetch.getProveedores(this.fecthForm.value).subscribe(response => {
@@ -93,25 +103,27 @@ export class ProveedorAgenciaComponent {
 
   protected newProvider(){
     this._fetch.newProveedor(this.newItemForm.value).subscribe(response =>{
-    this.getProveedores('1');
+    this.getProveedores(1);
     })
   }
   protected modifyProvider(){
     this._fetch.changeProveedor(this.changeItemForm.value,this.historial).subscribe(response =>{
-      //TODO:condicion
       this.modalRef.dismiss('close');
-      this.getProveedores('1');
+      this.getProveedores(1);
     })
   }
   protected deleteProvider(){
     this.DeleteItemForm.patchValue({ neo_id: this.editName[1]});
-    //todo swal
-    this._fetch.deleteProveedor(this.DeleteItemForm.value).subscribe(response => {
-      //todo:swal
-      console.log("succes deletion")
-      this.getProveedores('1');
+    this._swal.swalConfirmationRequest(this._translate.instant('LINKS.ALERT_TITLE_DELETE'),this._translate.instant(" "), " ").then(
+      (res:any)=>{
+        if (res.isConfirmed) {
+          this._fetch.deleteProveedor(this.DeleteItemForm.value).subscribe(response => {
+            this._swal.swalSucces(this._translate.instant('LINKS.ALERT_RESPONSE1'),this._translate.instant(' '));
+            this.getProveedores(1);
+          });
+        }
     });
-  }
+  };
 
   open(content : any) {
     this.historial = this.editName[0];
@@ -120,9 +132,6 @@ export class ProveedorAgenciaComponent {
       windowClass: 'modal-element',
       size: "sm",
       modalDialogClass:"rounded-0" });
-  }
-
-  download(){
   }
 
 
@@ -137,4 +146,19 @@ export class ProveedorAgenciaComponent {
       return fetchInfo === this.fecthForm.get('datos_paginacion.tipo_orden')?.value;
   }
 
+  protected changeOfPage(){
+    this.getProveedores(this.currentPage);
+
+  }
+  public saveDataInCSV(name: string, data: Array<any>): void {
+    let csvContent = this._csvService.saveDataInCSV(data);
+
+    var hiddenElement = document.createElement('a');
+    hiddenElement.href = 'data:text/csv;charset=utf-8,' + encodeURI(csvContent);
+    hiddenElement.target = '_blank';
+    hiddenElement.download = name + '.csv';
+    hiddenElement.click();
+  }
+
 }
+
